+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1-wip
 
 - Added topics.
+- Fixed metadata example.
 
 ## 0.1.0